get_first_and_last_day_of_month() has been removed with unit tests. print_calculation() has been moved to helpers.py
