+++ src/helpers.py
@@ -1,45 +1,13 @@
 """This module provides helper functions that are commonly used across different parts of the application."""
 
-from calendar import monthrange
-from datetime import date
 from datetime import datetime
 from math import ceil
 from os import environ
-from typing import Union, Tuple
+import logging
 
 
-def get_first_and_last_day_of_month(
-    month: Union[str, int, None] = None, year: Union[str, int, None] = None
-) -> Tuple[str, str]:
-    """Calculate the first and last day of the specified month and year, or the current month and year if not specified.
-
-    :param month: Optional; an integer (1-12) or a string representing the month. If not specified, the current month is used.
-    :param year: Optional; an integer representing the year or a string that can be converted to an integer. If not specified, the current year is used.
-    :return: A tuple containing two strings, the first and last day of the specified or current month and year in 'YYYY-MM-DD' format.
-    """
-    today = date.today()
-    if year is None:
-        year = today.year
-    else:
-        if isinstance(year, str) or isinstance(year, int):
-            year = int(year)
-        else:
-            raise ValueError("Year must be a string or integer that can be converted to an integer")
-
-    if month is None:
-        month = today.month
-    else:
-        if isinstance(month, str) or isinstance(month, int):
-            month = int(month)
-        else:
-            raise ValueError("Month must be a string or integer that can be converted to an integer")
-
-    if not 1 <= month <= 12:
-        raise ValueError("Month must be between 1 and 12")
-
-    first_day = date(year, month, 1)
-    last_day = date(year, month, monthrange(year, month)[1])
-    return first_day.strftime("%Y-%m-%d"), last_day.strftime("%Y-%m-%d")
+logger = logging.getLogger(__name__)
+logging.basicConfig(format="%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)", level=logging.INFO)
 
 
 def calculate_interval_hours(start_time, end_time) -> int:
@@ -70,3 +38,20 @@ def calculate_interval_hours(start_time, end_time) -> int:
         total_hours += ceil(delta.seconds / 3600)
 
     return total_hours
+
+
+def print_calculation(first_name: str, presenting: dict) -> None:
+    """Calculate and log the total days, meal payments, hours, and hour payments for a child in kindergarten.
+
+    :param first_name: The first name of the child.
+    :param presenting: Dictionary with actual start_time and end_time for each day.
+    """
+    hours = sum([(calculate_interval_hours(i["start_time"], i["end_time"])) for i in presenting])
+
+    logger.info(f"{first_name}. Days in kindergarten: '{len(presenting)}'.")
+    rounded_meals_payment = round(len(presenting) * int(environ["PAYMENT_FOR_MEALS_FOR_DAY"]), 2)
+    logger.info(f"{first_name}. Payment for meals: '{'{:.2f}'.format(rounded_meals_payment)}' zl.")
+
+    logger.info(f"{first_name}. Hours in kindergarten: '{hours}'.")
+    rounded_hours_payment = round(hours * float(environ["PAYMENT_FOR_HOUR"]), 2)
+    logger.info(f"{first_name}. Payment for hours: '{'{:.2f}'.format(rounded_hours_payment)}' zl.")
